Fix Integration Tests that were assuming 2 default core connections or depending on onUp for a new node

+++ driver-core/src/test/java/com/datastax/driver/core/StateListenerTest.java
@@ -29,7 +29,7 @@ import static com.datastax.driver.core.TestUtils.*;
  */
 public class StateListenerTest {
 
-    @Test(groups = "short")
+    @Test(groups = "long")
     public void listenerTest() throws Throwable {
 
         CCMBridge.CCMCluster c = CCMBridge.buildCluster(1, Cluster.builder());
